# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'ingot-forms API'
  description: 'Forms management backend. A form is a single fillable document: one immutable definition plus one data set with a required expire date. Lifecycle: empty → draft (repeatable lenient saves) → confirmed (strict validation, locked forever). Expired forms answer 410 everywhere. Every error response is an RFC 9457 `application/problem+json` document.'
  license:
    name: MIT
    identifier: MIT
  version: 0.2.0
servers:
  -
    url: /
paths:
...
  '/api/manage/form-templates/{template}/presentations':
    get:
      summary: "List a template's presentations"
      description: 'The other history, numbered apart from the definitions — a relabelled option is not a new model, and the two counts say so.'
      operationId: listTemplatePresentations
      parameters:
        -
          name: template
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
      responses:
        '200':
          description: 'The presentation history.'
        '404':
          $ref: '#/components/responses/FormTemplateNotFound'
    post:
      summary: 'Publish a presentation into a template'
      description: 'The answer is the number it was published as. It is judged against the definition currently in use — a presentation showing an item that definition does not declare is refused here, with the findings pointing at the item — and nothing about what is in use changes.'
      operationId: publishTemplatePresentation
      parameters:
        -
          name: template
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishPresentationRequest'
      responses:
        '201':
          description: 'Published. Nothing in use has changed.'
          content:
            application/json:
              schema:
                properties:
                  presentation: { type: integer, example: 4 }
                type: object
        '404':
          $ref: '#/components/responses/FormTemplateNotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          description: 'The presentation breaks its meta-schema, or does not fit the definition currently in use.'
...
  '/api/manage/form-templates/{template}':
    get:
      summary: 'Read a form template'
      description: 'The pair in use is given as version numbers, which is what `PUT …/current` takes and what the two history addresses read. `forms` counts what is made of any version this template has ever published — the number a delete is refused over.'
      operationId: readFormTemplate
      parameters:
        -
          name: template
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
      responses:
        '200':
          description: 'The template.'
          content:
            application/json:
              schema:
                properties:
                  id: { type: string, format: uuid }
                  name: { type: string }
                  createdAt: { type: string, format: date-time }
                  createdBy: { type: [string, 'null'] }
                  definition: { description: 'The definition version new forms are made of.', type: integer }
                  presentation: { description: 'The presentation version that shows them, or null when this template shows nothing.', type: [integer, 'null'] }
                  forms: { description: 'How many forms are made of any version this template ever published.', type: integer }
                type: object
        '404':
          $ref: '#/components/responses/FormTemplateNotFound'
...
  '/api/manage/form-templates/{template}/definitions/{seq}':
    get:
      summary: 'Read one published definition'
      description: 'The document as it was accepted, byte for byte.'
      operationId: readTemplateDefinition
      parameters:
        -
          name: template
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
        -
          name: seq
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9]+'
      responses:
        '200':
          description: 'The definition.'
        '404':
          $ref: '#/components/responses/TemplateVersionNotFound'
  '/api/manage/form-templates/{template}/presentations/{seq}':
    get:
      summary: 'Read one published presentation'
      description: 'The document as it was accepted, byte for byte.'
      operationId: readTemplatePresentation
      parameters:
        -
          name: template
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
        -
          name: seq
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9]+'
      responses:
        '200':
          description: 'The presentation.'
        '404':
          $ref: '#/components/responses/TemplateVersionNotFound'
...
  '/api/manage/form-templates/{template}/name':
    put:
      summary: 'Rename a form template'
      description: 'The label only. A name is never an identifier here — nothing looks a template up by it — so this changes nothing about what forms made from this template ask or how they are shown.'
      operationId: renameFormTemplate
      parameters:
        -
          name: template
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameTemplateRequest'
      responses:
        '204':
          description: Renamed.
        '404':
          $ref: '#/components/responses/FormTemplateNotFound'
        '422':
          description: 'The name is blank or longer than the column keeps.'
...
components:
  schemas:
...
    ProblemError:
      description: "One validation error, mapped 1:1 from the engine's report or a constraint violation."
      required:
        - pointer
        - code
        - message
      properties:
        pointer:
          description: 'RFC 6901 JSON Pointer into the offending document ("" is its root).'
          type: string
        code:
          description: 'Machine-readable code, e.g. `schema.required`, `request.range` or `form.field.duplicate-name`.'
          type: string
        message:
          type: string
        input:
          description: 'The offending value — echoed only when it is a scalar.'
          type:
            - string
            - number
            - boolean
      type: object
      additionalProperties: false
    Problem:
      description: 'RFC 9457 problem details. `type` is a URN `urn:problem:ingot-forms:<slug>`; validation problems carry the `errors` extension member.'
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ProblemError'
      type: object
      additionalProperties: false
...
    PublishPresentationRequest:
      required:
        - presentation
      properties:
        presentation:
          description: 'The presentation to publish, per `GET /api/schemas/presentation`. Judged against the definition currently in use; a presentation showing an item that definition does not declare is refused here, with the findings pointing at the item.'
          minProperties: 1
          type: object
          example:
            engine: core-html
            items:
              -
                name: email
                widget: text
              -
                widget: confirm
      type: object
      additionalProperties: false
...
    RenameTemplateRequest:
      required:
        - name
      properties:
        name:
          description: 'What to call this template from now on. A label and never an identifier, so changing it changes nothing else.'
          type: string
          example: 'Damage report (2027)'
          maxLength: 255
      type: object
      additionalProperties: false
...
  responses:
...
    UnsupportedMediaType:
      description: 'The request body is not `application/json` — no other media type is accepted.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            unsupported-media-type:
              summary: 'A form-encoded or missing content type'
              value:
                type: 'urn:problem:ingot-forms:unsupported-media-type'
                title: 'Only application/json request bodies are accepted.'
                status: 415
...
    FormTemplateNotFound:
      description: 'No form template with this id.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            form-template-not-found:
              summary: 'Unknown id'
              value:
                type: 'urn:problem:ingot-forms:form-template-not-found'
                title: 'Form template not found.'
                status: 404
    TemplateVersionNotFound:
      description: 'This template published no such version. A number is a place in one of its two histories, and an unknown one is a missing document rather than a bad request.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            template-version-not-found:
              summary: 'A number this template never handed out'
              value:
                type: 'urn:problem:ingot-forms:template-version-not-found'
                title: 'The form template has no such version.'
                status: 404
...